make: Style cleanups and target cleanups
Explicitly defining the target for the build is important,
but it's nice to be able to assume defaults based upon the
host system, since often it is useful to use tu58fs from
a laptop rather than a Raspberry Pi or a Beaglebone.

Additionally, this cleans up a bunch of tabs in places
where make can misconstrue as misplaced rules, adjusts
variables to be more standard (CCFLAGS instead of CFLAGS)
to prevent make mis-steps, as well as using pattern rules
to help make the build directory and simplify the
rules considerably.