macos: Simplify some of the platform dependent bits
On older versions of macos prior to Sonoma (14.5), clock_gettime
was missing from the standard libraries. On versions after this,
Apple rectified the situation. Unfortunately, probing for it
is a bit of a pain, so this creates a config.h file we can use
to help tune the source a bit for the platforms it is built on.

Since modern macos versions include this function, the default
for the config is to assume that clock_gettime is available --
especially given that macos is quite agressive at updating its
system, this should be a reasonable default.